activity-service/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==2.6.1
